+++ controllers/users.js
@@ -29,7 +29,7 @@ const {
 const { addLog } = require("../models/logs");
 const { getUserStatus } = require("../models/userStatus");
 const config = require("config");
-const { generateUniqueUsername } = require("../services/users");
+const { generateUniqueUsername, getUsersWithIncompleteTasks } = require("../services/users");
 const discordDeveloperRoleId = config.get("discordDeveloperRoleId");
 
 const verifyUser = async (req, res) => {
@@ -1031,7 +1031,7 @@ const updateUsernames = async (req, res) => {
 
 const getUsersWithAbandonedTasks = async (req, res) => {
   try {
-    const data = await userQuery.fetchUsersWithAbandonedTasks();
+    const data = await getUsersWithIncompleteTasks();
     return res.status(200).json({ message: "Users with abandoned tasks fetched successfully", data });
   } catch (error) {
     logger.error("Error in getting user who abandoned tasks:", error);
