+++ services/users.js
@@ -1,44 +1,29 @@
-const { USERS_PATCH_HANDLER_SUCCESS_MESSAGES, USERS_PATCH_HANDLER_ERROR_MESSAGES } = require("../constants/users");
 const firestore = require("../utils/firestore");
 const { formatUsername } = require("../utils/username");
 const userModel = firestore.collection("users");
-const archiveUsers = async (usersData) => {
-  const batch = firestore.batch();
-  const usersBatch = [];
-  const summary = {
-    totalUsersArchived: 0,
-    totalOperationsFailed: 0,
-    updatedUserDetails: [],
-    failedUserDetails: [],
-  };
+const { fetchUsersNotInDiscordServer } = require("../models/users");
+const { fetchIncompleteTaskForUser } = require("../models/tasks");
 
-  usersData.forEach((user) => {
-    const { id, first_name: firstName, last_name: lastName } = user;
-    const updatedUserData = {
-      ...user,
-      roles: {
-        ...user.roles,
-        archived: true,
-      },
-      updated_at: Date.now(),
-    };
-    batch.update(userModel.doc(id), updatedUserData);
-    usersBatch.push({ id, firstName, lastName });
-  });
-
+const getUsersWithIncompleteTasks = async () => {
   try {
-    await batch.commit();
-    summary.totalUsersArchived += usersData.length;
-    summary.updatedUserDetails = [...usersBatch];
-    return {
-      message: USERS_PATCH_HANDLER_SUCCESS_MESSAGES.ARCHIVE_USERS.SUCCESSFULLY_COMPLETED_BATCH_UPDATES,
-      ...summary,
-    };
-  } catch (err) {
-    logger.error("Firebase batch Operation Failed!");
-    summary.totalOperationsFailed += usersData.length;
-    summary.failedUserDetails = [...usersBatch];
-    return { message: USERS_PATCH_HANDLER_ERROR_MESSAGES.ARCHIVE_USERS.BATCH_DATA_UPDATED_FAILED, ...summary };
+    const eligibleUsersWithTasks = [];
+
+    const userSnapshot = await fetchUsersNotInDiscordServer();
+
+    for (const userDoc of userSnapshot.docs) {
+      const user = userDoc.data();
+
+      // Check if the user has any tasks with status not in [Done, Complete]
+      const abandonedTasksQuerySnapshot = await fetchIncompleteTaskForUser(user);
+
+      if (!abandonedTasksQuerySnapshot.empty) {
+        eligibleUsersWithTasks.push(user);
+      }
+    }
+    return eligibleUsersWithTasks;
+  } catch (error) {
+    logger.error(`Error in getting users who abandoned tasks:  ${error}`);
+    throw error;
   }
 };
 
@@ -63,6 +48,6 @@ const generateUniqueUsername = async (firstName, lastName) => {
 };
 
 module.exports = {
-  archiveUsers,
   generateUniqueUsername,
+  getUsersWithIncompleteTasks,
 };
