+++ test/fixtures/tasks/tasks.js
@@ -140,89 +140,5 @@ module.exports = () => {
       createdAt: 1644753600,
       updatedAt: 1644753600,
     },
-    {
-      id: "task1_id",
-      title: "Abandoned Task 1",
-      type: "feature",
-      status: "IN_PROGRESS",
-      priority: "HIGH",
-      percentCompleted: 50,
-      createdAt: 1727027666,
-      updatedAt: 1727027999,
-      startedOn: 1727027777,
-      endsOn: 1731542400,
-      assignee: "archived_user1",
-      assigneeId: "user1_id",
-      github: {
-        issue: {
-          html_url: "https://github.com/org/repo/issues/1",
-          url: "https://api.github.com/repos/org/repo/issues/1",
-        },
-      },
-      dependsOn: [],
-    },
-    {
-      id: "task2_id",
-      title: "Abandoned Task 2",
-      type: "bug",
-      status: "BLOCKED",
-      priority: "MEDIUM",
-      percentCompleted: 30,
-      createdAt: 1727027666,
-      updatedAt: 1727027999,
-      startedOn: 1727027777,
-      endsOn: 1731542400,
-      assignee: "archived_user2",
-      assigneeId: "user2_id",
-      github: {
-        issue: {
-          html_url: "https://github.com/org/repo/issues/2",
-          url: "https://api.github.com/repos/org/repo/issues/2",
-        },
-      },
-      dependsOn: [],
-    },
-    {
-      id: "task3_id",
-      title: "Completed Archived Task",
-      type: "feature",
-      status: "DONE",
-      priority: "LOW",
-      percentCompleted: 100,
-      createdAt: 1727027666,
-      updatedAt: 1727027999,
-      startedOn: 1727027777,
-      endsOn: 1731542400,
-      assignee: "archived_user1",
-      assigneeId: "user1_id",
-      github: {
-        issue: {
-          html_url: "https://github.com/org/repo/issues/3",
-          url: "https://api.github.com/repos/org/repo/issues/3",
-        },
-      },
-      dependsOn: [],
-    },
-    {
-      id: "task4_id",
-      title: "Active User Task",
-      type: "feature",
-      status: "IN_PROGRESS",
-      priority: "HIGH",
-      percentCompleted: 75,
-      createdAt: 1727027666,
-      updatedAt: 1727027999,
-      startedOn: 1727027777,
-      endsOn: 1731542400,
-      assignee: "active_user",
-      assigneeId: "user3_id",
-      github: {
-        issue: {
-          html_url: "https://github.com/org/repo/issues/4",
-          url: "https://api.github.com/repos/org/repo/issues/4",
-        },
-      },
-      dependsOn: [],
-    },
   ];
 };
